HeatFix: check if the old patch's fps limit results in missed Heat gain I did suspect that it was a possibility but this test confirms it.
Contrary to what I believed at the time, several sources of Heat gain
(e.g. hitting the enemy) seem to get applied on a single frame instead
of increasing Heat over multiple frames. So if we happen to skip the
HeatUpdate function on a frame where such a gain was supposed to happen,
we'll miss out on the Heat gain for that particular attack.